refactor: address PR review comments for HNSW eviction tests
Documentation (execplan):
- Define HNSW acronym (Hierarchical Navigable Small World) in title
- Define ExecPlan term on first use
- Change "Artifacts" to "Artefacts" (en-GB spelling)
- Expand ADR acronym (Architecture Decision Record)
- Add missing comma after "tests pass"
- Update roadmap snippet to show [x] (completed)
- Tick validation checkboxes to reflect completed state

Code changes (kani_proofs.rs):
- Change EdgeAssertion fields from u32 to usize to match graph API
- Remove redundant `as usize` casts in assert_node_link/assert_no_node_link

Code changes (commit/tests.rs):
- Strengthen healing assertion to verify node 2 connects to entry node 0
- Extract HealingTestContext struct for base layer healing tests
- Add benign_deferred_scrub_is_noop_when_edge_already_removed test

Code changes (functional_ari_nmi.rs):
- Document threshold difference between iris (0.65) and ruspini (0.95)
- Explain iris's relaxed thresholds due to overlapping class boundaries

+++ chutoro-core/src/hnsw/kani_proofs.rs
@@ -314,14 +314,14 @@ fn setup_eviction_test_graph(params: HnswParams) -> Graph {
 
 #[cfg(kani)]
 struct EdgeAssertion {
-    source: u32,
-    target: u32,
+    source: usize,
+    target: usize,
     level: usize,
 }
 
 #[cfg(kani)]
 impl EdgeAssertion {
-    fn new(source: u32, target: u32, level: usize) -> Self {
+    fn new(source: usize, target: usize, level: usize) -> Self {
         Self {
             source,
             target,
@@ -334,8 +334,8 @@ impl EdgeAssertion {
 #[cfg(kani)]
 fn assert_node_link(graph: &Graph, edge: EdgeAssertion, message: &str) {
     let has_link = graph
-        .node(edge.source as usize)
-        .map(|n| n.neighbours(edge.level).contains(&(edge.target as usize)))
+        .node(edge.source)
+        .map(|n| n.neighbours(edge.level).contains(&edge.target))
         .unwrap_or(false);
     kani::assert(has_link, message);
 }
@@ -344,8 +344,8 @@ fn assert_node_link(graph: &Graph, edge: EdgeAssertion, message: &str) {
 #[cfg(kani)]
 fn assert_no_node_link(graph: &Graph, edge: EdgeAssertion, message: &str) {
     let has_link = graph
-        .node(edge.source as usize)
-        .map(|n| n.neighbours(edge.level).contains(&(edge.target as usize)))
+        .node(edge.source)
+        .map(|n| n.neighbours(edge.level).contains(&edge.target))
         .unwrap_or(false);
     kani::assert(!has_link, message);
 }
